added vars to dict; fixed long lines and var names; moved wrapper registry
