+++ hive/envs/base.py
@@ -1,8 +1,6 @@
 from abc import ABC, abstractmethod
-from typing import List
-import gym
 
-from hive.utils.registry import registry, Registrable
+from hive.utils.registry import Registrable
 
 
 class BaseEnv(ABC, Registrable):
@@ -160,44 +158,3 @@ class ParallelEnv(BaseEnv):
             self._turn,
             self._info,
         )
-
-
-class EnvWrapper(Registrable):
-    """A wrapper for callables that produce environment wrappers.
-
-    These wrapped callables can be partially initialized through configuration
-    files or command line arguments.
-    """
-
-    @classmethod
-    def type_name(cls):
-        """
-        Returns:
-            "init_fn"
-        """
-        return "init_fn"
-
-
-class CompositeEnvWrapper(gym.Wrapper):
-    """This Logger aggregates multiple env wrappers."""
-
-    def __init__(self, env, wrapper_list: List[EnvWrapper]):
-        for wrapper in wrapper_list:
-            env = wrapper(env)
-        super().__init__(env)
-
-
-registry.register_all(
-    EnvWrapper,
-    {
-        "RecordEpisodeStatistics": gym.wrappers.RecordEpisodeStatistics,
-        "ClipAction": gym.wrappers.ClipAction,
-        "NormalizeObservation": gym.wrappers.NormalizeObservation,
-        "TransformObserva": gym.wrappers.TransformObservation,
-        "NormalizeReward": gym.wrappers.NormalizeReward,
-        "TransformReward": gym.wrappers.TransformReward,
-        "CompositeEnvWrapper": CompositeEnvWrapper,
-    },
-)
-
-get_wrapper = getattr(registry, f"get_{EnvWrapper.type_name()}")
